+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii2 multiple input change log
 
 - Renamed `limit` option to `max`
 - Changed namespace from `unclead\widgets` to `yii\multipleinout`
+- #92: Adjustments for correct work with AR relations
 
 1.4.1
 =====